:recycle: gamble 페이지 분리
