+++ src/@views/Home/index.tsx
@@ -1,0 +1,7 @@
+import {FC} from 'react';
+
+const HomeScreen: FC = () => {
+  return <>Hello, World</>;
+};
+
+export default HomeScreen;
